requirements.txt
requests==2.28.2
mysql-connector-python==8.0.32
pytest==7.2.1
python-dotenv==0.21.1